Fix ACLs and cleanup operations.
The default ACLs are figured out and added to the working volume directory
before the job steps are run, that way any files that are created by a
step are accessible to every other step, regardless of what user was
used in the each container.

We were skipping over clean up operations when a step failed, which was
leaking ingresses.